requirements.txt
Flask==2.3.2
Flask-MySQLdb==1.0.1
pandas==2.0.3
scikit-learn==1.2.2
numpy==1.24.3
Jinja2==3.1.2
Werkzeug==2.3.6
click==8.1.3
itsdangerous==2.1.2
MarkupSafe==2.1.3
flask_wtf
mysql-connector-python
openpyxl
email_validator
